update app with instagram and daily refresh

+++ pull_instagram.py
@@ -5,7 +5,7 @@ import pandas as pd
 from apify_client import ApifyClient
 # Replace with your actual Apify API Key
 from config import APIFY_API_KEY, TASK_ID
-
+import os
 from apify_client import ApifyClient
 import pandas as pd
 
@@ -29,9 +29,17 @@ def fetch_all_fields():
 
     return df
 
-# Fetch latest posts and return as DataFrame
-df_posts = fetch_all_fields()
 
-# Save to CSV (modify path as needed)
+# Construct the filename for today's date
 new_data_file = 'data/new_insta_data/new_insta_data' + datetime.now().strftime('%d%m%Y') + '.csv'
-df_posts.to_csv(new_data_file, index=False)
+
+# Check if the file already exists
+if not os.path.exists(new_data_file):
+    print(f"Fetching new Instagram data and saving to {new_data_file}")
+    
+    df_posts = fetch_all_fields()  # Fetch latest posts
+    
+    # Save to CSV
+    df_posts.to_csv(new_data_file, index=False)
+else:
+    print(f"File already exists: {new_data_file}. Skipping fetch.")
